refactor: pass to fetch() only what is needed
