+++ src/api/install.ts
@@ -8,7 +8,7 @@ import cloneDeep = require('lodash.clonedeep')
 import {PnpmOptions, StrictPnpmOptions, Dependencies} from '../types'
 import createGot from '../network/got'
 import getContext, {PnpmContext} from './getContext'
-import installMultiple from '../install/installMultiple'
+import installMultiple, {InstalledPackage} from '../install/installMultiple'
 import save from '../save'
 import linkPeers, {linkPeersWhenSymlinksPreserved} from '../install/linkPeers'
 import runtimeError from '../runtimeError'
@@ -16,7 +16,6 @@ import getSaveType from '../getSaveType'
 import {sync as runScriptSync} from '../runScript'
 import postInstall from '../install/postInstall'
 import extendOptions from './extendOptions'
-import {InstalledPackage} from '../install/fetch'
 import pnpmPkgJson from '../pnpmPkgJson'
 import lock from './lock'
 import {save as saveStore, Store} from '../fs/storeController'
